requirements.txt
absl-py==0.5.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
docopt==0.6.2
enum34==1.1.6
future==0.16.0
idna==2.7
imageio==2.4.1
jsonpickle==1.2
kiwisolver==1.0.1
matplotlib==3.0.0
mock==2.0.0
more-itertools==4.3.0
mpyq==0.2.5
munch==2.3.2
numpy==1.15.2
pathlib2==2.3.2
pbr==4.3.0
Pillow==5.3.0
pluggy==0.7.1
portpicker==1.2.0
probscale==0.2.3
protobuf==3.6.1
py==1.6.0
pygame==1.9.4
pyparsing==2.2.2
pysc2==3.0.0
pytest==3.8.2
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
s2clientprotocol==4.10.1.75800.0
sacred==0.8.1
scipy==1.1.0
six==1.11.0
sk-video==1.1.10
snakeviz==1.0.0
tensorboard-logger==0.1.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1.1
urllib3==1.23
websocket-client==0.53.0
whichcraft==0.5.2
wrapt==1.10.11
git+https://github.com/oxwhirl/smac.git
